feat :更新属性candleSolid --> candleHollow 添加属性macdHollow ,去除所有枚类型,分离Y轴valueFormater与MainValueFormatter 需要分别设置,修改指标图valueFromatter不起作用的问题

+++ klinelib/src/main/java/com/icechao/klinelib/formatter/IValueFormatter.java
@@ -17,5 +17,5 @@ public interface IValueFormatter {
      * @param value 传入的value值
      * @return 返回字符串
      */
-    String format(float value);
+    String format(double value);
 }
